+++ Runtime/Component/Legend/Legend.cs
@@ -80,8 +80,8 @@ namespace XCharts.Runtime
         [SerializeField] private List<string> m_Data = new List<string>();
         [SerializeField] private List<Sprite> m_Icons = new List<Sprite>();
         [SerializeField] private List<Color> m_Colors = new List<Color>();
-        [SerializeField] protected ImageStyle m_Background = new ImageStyle() { show = false };
-        [SerializeField] protected Padding m_Padding = new Padding();
+        [SerializeField][Since("v3.1.0")] protected ImageStyle m_Background = new ImageStyle() { show = false };
+        [SerializeField][Since("v3.1.0")] protected Padding m_Padding = new Padding();
 
         public LegendContext context = new LegendContext();